Made module for layout tests
